master/website/holons/requirements.txt
3to2==1.1.1
alabaster==0.7.12
amqp==2.5.2
anyascii==0.1.7
apiclient==1.0.4
appdirs==1.4.4
applib==1.2
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.2.0
Automat==20.2.0
awscli==1.16.174
Babel==2.9.0
backcall==0.2.0
beautifulsoup4==4.8.2
billiard==3.6.1.0
bleach==3.3.0
blis==0.4.1
boto3==1.9.102
botocore==1.12.164
cachetools==4.1.0
catalogue==1.0.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cmdln==2.0.0
CodeIntel==2.0.0
colorama==0.3.9
constantly==15.1.0
convertdate==2.3.1
coverage==5.5
cryptography==2.9.2
cssselect==1.1.0
cymem==2.0.4
Cython==0.29.23
decorator==4.4.2
defusedxml==0.7.1
Django==2.2
django-appconf==1.0.4
django-bulk-update==2.2.0
django-cors-headers==3.2.1
django-elasticsearch-dsl==0.5.1
django-extensions==3.1.2
django-filter==2.4.0
django-hashid-field==3.2.1
django-imagefit==0.6.2
django-jsonfield-backport==1.0.3
django-mailbox==4.8.1
django-modelcluster==5.1
django-money==1.3.1
django-seed==0.2.2
django-taggit==1.3.0
django-timezone-field==4.1.1
django-treebeard==4.4
djangorestframework==3.12.2
docopt==0.6.2
docutils==0.14
draftjs-exporter==2.1.7
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
entrypoints==0.3
esprima==4.0.1
et-xmlfile==1.0.1
Faker==8.0.0
feedgen==0.9.0
feedparser==6.0.2
future==0.18.2
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.15.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.4
hashids==1.3.1
hijri-converter==2.1.1
holidays==0.10.5.2
html5lib==1.1
httplib2==0.18.1
hyperlink==20.0.1
icalendar==4.0.7
idna==2.8
imagesize==1.2.0
IMAPClient==2.1.0
importlib-metadata==1.3.0
incremental==17.5.0
Inflector==3.0.1
ipaddress==1.0.22
ipykernel==5.5.3
# ipython==7.22.0
# ipython-genutils==0.2.0
ipywidgets==7.6.3
itemadapter==0.1.1
itemloaders==1.0.3
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kombu==4.6.7
korean-lunar-calendar==0.2.1
l18n==2020.6.1
libclang-py3==3.9.0
libmagic==1.0
logger==1.4
ls.joyful==0.1.3
ls.joyous==1.4.0
lxml==4.5.2
Markdown==3.0.1
MarkupSafe==1.1.1
mistune==0.8.4
model-bakery==1.2.1
more-itertools==8.0.2
murmurhash==1.0.2
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
num2words==0.5.10
numpy
oauthlib==3.1.0
openpyxl==3.0.6
packaging==20.9
pandas==0.25.3
pandocfilters==1.4.3
parsel==1.6.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.10.1
prompt-toolkit==3.0.18
Protego==0.1.16
protobuf==3.12.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
ptyprocess==0.7.0
py-moneyed==0.8.0
py-trello==0.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
Pygments==2.7.4
PyHamcrest==2.0.2
pyhunter==0.3
PyMeeus==0.4.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest-runner==4.4
python-dateutil==2.8.1
python-http-client==3.3.1
python-magic==0.4.22
python-telegram-bot==12.6.1
pytz==2021.1
PyYAML==3.13
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
queuelib==1.5.0
redis==3.4.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer
Scrapy==2.5.0
Send2Trash==1.5.0
sendgrid==6.4.8
sentry-sdk==0.19.5
service-identity==18.1.0
sgmllib3k==1.0.0
six==1.12.0
snowballstemmer==2.1.0
soupsieve==1.8
spacy==2.2.4
Sphinx==3.4.3
sphinx-autodoc-annotation==1.0.post1
sphinx-rtd-theme==0.5.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-django==0.5.1
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
srsly==1.0.2
starkbank-ecdsa==1.1.0
stripe==2.57.0
tablib==3.0.0
terminado==0.9.4
testpath==0.4.4
text-unidecode==1.3
thinc==7.4.0
tld==0.9.3
tornado==6.1
tqdm==4.45.0
traitlets==5.0.5
Twisted==20.3.0
uritemplate==3.0.1
urllib3==1.24.3
vine==1.3.0
virtualenv==16.7.7
w3lib==1.22.0
wagtail==2.12.2
wasabi==0.6.0
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
zipp==0.6.0
zope.cachedescriptors==4.3.1
zope.interface==5.1.2
